Visal servoing is alive
