+++ irob_subtask_logic/launch/camera_full.launch
@@ -1,0 +1,17 @@
+<!-- Launch file for vessel cutting training excercise -->
+
+<launch>
+
+
+
+    <include file="$(find irob_motion)/launch/surgeme_server.launch"/>
+    <include file="$(find irob_subtask_logic)/launch/camera.launch"/>
+    <include file="$(find irob_vision_support)/launch/stereo_cam_blackmagic_raw.launch"/>
+    <include file="$(find irob_vision_support)/launch/aruco_detector.launch"/>
+    <include file="$(find irob_vision_support)/launch/marker_displacement.launch"/>
+
+
+
+
+
+</launch>
